% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\docType{data}
\name{attributes}
\alias{attributes}
\title{Workplace employees}
\format{
Dataframe with 232 rows and 3 columns.
\describe{
 \item{id}{Numeric id for the employee}
 \item{time}{Time when the covariate value changes}
 \item{department}{Factor variable with the department of the employee}
}
}
\source{
\url{http://www.sociopatterns.org/datasets/test/}
}
\usage{
data(attributes)
}
\description{
Employees that interact in the `edges` data object. Contains information
on the department they work in.
}
\references{
Genois & Barrat (2018) EPJ Data Science 7:11 \url{https://doi.org/10.1140/epjds/s13688-018-0140-1}
}
\keyword{datasets}
