% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\docType{data}
\name{constant}
\alias{constant}
\title{constant}
\format{
A list of length 200 with each element a data.frame with 10000 rows
and 3 columns that contains a generated REH:
\describe{
 \item{time}{numeric value indicating the time of the event, since onset
 of the observation period}
 \item{sender}{numeric value with the id of the sender of the event}
 \item{receiver}{numeric value with the id of the receiver of the event}
}
}
\usage{
constant
}
\description{
200 generated datasets in the "constant" scenario.
}
\keyword{datasets}
