% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar.R
\docType{data}
\name{covar}
\alias{covar}
\title{covar}
\format{
A dataframe with 30 rows and 3 variables:
\describe{
 \item{id}{numeric id for the actor}
 \item{time}{numeric value that indicates when the covariate value changes}
 \item{z}{attribute value sampled from the standard normal distribution}
}
}
\usage{
covar
}
\description{
Attributes for the 30 actors that interact in the generated datasets.
}
\keyword{datasets}
