% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createwindows.R
\name{createwindows}
\alias{createwindows}
\title{createwindows}
\usage{
createwindows(length, end, start = NULL, overlap = 2/3)
}
\arguments{
\item{length}{numeric value that indicates the length of the window in the
correct time units}

\item{end}{numeric value that indicates the maximum time for the windows}

\item{start}{numeric value that indicates the minimum time for the windows, 
if not supplied it is assumed to be 0}

\item{overlap}{numeric value between 0 and 1 indicating the proportion of 
overlap between subsequent windows}
}
\description{
A function that creates a dataframe with window timepoints
}
