% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddfit_constant.R
\docType{data}
\name{ddfit_constant}
\alias{ddfit_constant}
\title{ddfit_constant}
\format{
A list of length 200, each element a list with the fits (remstimate)
objects for each window.
}
\usage{
ddfit_constant
}
\description{
Results for the moving window REM with data-driven window widths for the
200 generated datasets in the constant effects scenario.
}
\keyword{datasets}
