% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddfit_cyclic.R
\docType{data}
\name{ddfit_cyclic}
\alias{ddfit_cyclic}
\title{ddfit_cyclic}
\format{
A list of length 200, each element a list with the fits (remstimate)
objects for each window.
}
\usage{
ddfit_cyclic
}
\description{
Results for the moving window REM with data-driven window widths for the
200 generated datasets in the cyclic effects scenario.
}
\keyword{datasets}
