% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddfit_gradual.R
\docType{data}
\name{ddfit_gradual}
\alias{ddfit_gradual}
\title{ddfit_gradual}
\format{
A list of length 200, each element a list with the fits (remstimate)
objects for each window.
}
\usage{
ddfit_gradual
}
\description{
Results for the moving window REM with data-driven window widths for the
200 generated datasets in the gradual effects scenario.
}
\keyword{datasets}
