% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddfit_mixed.R
\docType{data}
\name{ddfit_mixed}
\alias{ddfit_mixed}
\title{ddfit_mixed}
\format{
A list of length 200, each element a list with the fits (remstimate)
objects for each window.
}
\usage{
ddfit_mixed
}
\description{
Results for the moving window REM with data-driven window widths for the
200 generated datasets in the mixed effects scenario.
}
\keyword{datasets}
