% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddwindows_gradual.R
\docType{data}
\name{ddwindows_gradual}
\alias{ddwindows_gradual}
\title{ddwindows_gradual}
\format{
A list of length 200, each element a list with two objects. One
with the data-driven window widths, and one with the evaluated options for 
the window widths at each timepoints.
}
\usage{
ddwindows_gradual
}
\description{
Data-driven windows for the 200 generated datasets in the gradual effects
scenario.
}
\keyword{datasets}
