% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddwindows_mixed.R
\docType{data}
\name{ddwindows_mixed}
\alias{ddwindows_mixed}
\title{ddwindows_mixed}
\format{
A list of length 200, each element a list with two objects. One
with the data-driven window widths, and one with the evaluated options for 
the window widths at each timepoints.
}
\usage{
ddwindows_mixed
}
\description{
Data-driven windows for the 200 generated datasets in the mixed effects
scenario.
}
\keyword{datasets}
