% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyneval.R
\name{dyneval}
\alias{dyneval}
\title{dyneval}
\usage{
dyneval(
  K,
  tie_effects,
  edgelist,
  covar,
  directed = FALSE,
  actors = NULL,
  ncores = 1
)
}
\arguments{
\item{K}{number of segments K to divide the edgelist in}

\item{tie_effects}{an object of class \code{"\link[stats]{formula}"} (or one 
that can be coerced to that class): a symbolic description of the effects in 
the tie-oriented model for which statistics are computed, see 'Details' for 
the available effects and their corresponding statistics}

\item{edgelist}{an object of class \code{"\link[base]{data.frame}"} or 
\code{"\link[base]{matrix}"} characterizing the relational event history 
sorted by time with columns `time`, `actor1`, `actor2` and optionally `type` 
and `weight`. Alternatively, an object of class \code{"\link[remify]{reh}"}}

\item{covar}{attributes dataframe with columns that describe, in order, 
actors' "id", attributes' change "time" and attributes' values.}

\item{directed}{logical value indicating whether dyadic events are directed (\code{TRUE}) or undirected (\code{FALSE}).}

\item{actors}{vector of actors that may be observed interacting in the network. If \code{NULL}, actor names will be taken from the input edgelist.}

\item{ncores}{number of threads for the parallelization (\code{default = 1}, no parallelization)}
}
\value{
List with elements "fit", "bfres" and "ngroup".
}
\description{
A function to test for time-varying network effects in a relational event
history.
}
