% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\docType{data}
\name{edges}
\alias{edges}
\title{Contacts in a workplace.}
\format{
Dataframe with 33751 rows and 4 columns that contains the relational event 
sequence with face-to-face contacts in the workplace. 
\describe{
 \item{time}{Numeric value that indicates the onset of the contact in 
seconds since onset of the study period.}
 \item{actor1}{Integer value with the ID of the first actor in the contact.}
 \item{actor2}{Integer value with the ID of the second actor in the contact.}
 \item{duration}{Numeric value that indicates the duration of the contact in 
seconds.}
}
}
\source{
\url{http://www.sociopatterns.org/datasets/test/}
}
\usage{
data(workplace2015)
}
\description{
This dataset contains the sequence of 33751 face-to-face contacts between 
employees measured in an office building in France, during two weeks in 
2015. The data collection involved 232 employees of 12 different 
departments. Wearable proximity sensors recorded with a 20-second time 
interval every face-to-face contact between individuals that was active in 
the last 20 second window. Following Genois & Barrat (2018), a contact 
between two individuals is defined as the set of successive time-windows of 
20 seconds during which the individuals are detected in contact. In the 
'edges' object, these records are collected in a relational event 
sequence with for every contact between a pair of individuals the time of 
onset and its duration. Time is expressed in seconds since the onset of the 
study period. The events in the sequence are undirected. Multiple events can 
occur at the same time point and the individuals can be in contact with 
multiple other individuals at the same time. Information on the department 
of the individuals is available in the 'attributes' data object.
}
\references{
Genois & Barrat (2018) EPJ Data Science 7:11 \url{https://doi.org/10.1140/epjds/s13688-018-0140-1}
}
\keyword{datasets}
