% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_ddwindows.R
\docType{data}
\name{example_ddwindows}
\alias{example_ddwindows}
\title{example_ddwindows}
\format{
List of length two:
\describe{
 \item{windows}{Dataframe with 343 rows and 6 columns. Contains the time for
 the begin and end of an optimal window around 343 time points. The 
 column 'm' refers to the number of events in the window. LogF to the 
 log fit, logC to the log complexity and logBF to the log Bayes factor
 for the Bayesian test for dynamic effects during the window.}
 \item{options}{List of length 343 with all the tried out window widths for 
 each time point and the results for the Bayesian test for dynamic effects 
 during the proposed window widths.}
}
}
\usage{
example_ddwindows
}
\description{
Contains the data-driven window widths for the empirical example data (see
`edges`).
}
\keyword{datasets}
