% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{generate}
\usage{
generate(formula, param, covar, M)
}
\arguments{
\item{formula}{formula object that can be supplied to the tie_effects
argument of remstats::remstats}

\item{param}{parameters for the effects}

\item{covar}{attributes dataframe with columns that describe, in order, 
actors' "id", attributes' change "time" and attributes' values.}

\item{M}{number of events}
}
\value{
dataframe with columns that describe subsequently the time, sender 
and receiver of the events
}
\description{
Function to generate REH data
}
