% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW1000_gradual.R
\docType{data}
\name{MW1000_gradual}
\alias{MW1000_gradual}
\title{MW1000_gradual}
\format{
A list of length 200, each element a list with the fits (remstimate)
objects for each window.
}
\usage{
MW1000_gradual
}
\description{
Results for the Moving Window REM with fixed width of 1000 time units for 
the 200 generated datasets in the gradual effects scenario.
}
\keyword{datasets}
