% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW2000_cyclic.R
\docType{data}
\name{MW2000_cyclic}
\alias{MW2000_cyclic}
\title{MW2000_cyclic}
\format{
A list of length 200, each element a list with the fits (remstimate)
objects for each window.
}
\usage{
MW2000_cyclic
}
\description{
Results for the Moving Window REM with fixed width of 2000 time units for 
the 200 generated datasets in the cyclic effects scenario.
}
\keyword{datasets}
