% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW4000_constant.R
\docType{data}
\name{MW4000_constant}
\alias{MW4000_constant}
\title{MW4000_constant}
\format{
A list of length 200, each element a list with the fits (remstimate)
objects for each window.
}
\usage{
MW4000_constant
}
\description{
Results for the Moving Window REM with fixed width of 4000 time units for 
the 200 generated datasets in the constant effects scenario.
}
\keyword{datasets}
