% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWrem.R
\name{MWrem}
\alias{MWrem}
\title{MWrem}
\usage{
MWrem(
  windows,
  edgelist,
  stats,
  actors,
  directed,
  method = "MLE",
  model = "tie",
  ncores = 1
)
}
\arguments{
\item{windows}{dataframe with two columns, the first column containing the 
time point for the start of each window and the second column containing
the time point for the end of each window}

\item{edgelist}{an object of class \code{"\link[base]{data.frame}"} or 
\code{"\link[base]{matrix}"} characterizing the relational event history sorted by 
time with columns `time`, `actor1`, `actor2` and optionally `type` and 
`weight`.}

\item{stats}{a \code{remstats} object: when \code{model="tie"}, \code{stats} is an array of statictis with dimensions \code{[M x D x P]}: where \code{M} is the number of events, \code{D} is the number of possible dyads (full riskset), \code{P} is the number of statistics; if \code{model="actor"}, \code{stats} is a list of two arrays named \code{rate} and \code{choice} with dimensions \code{[M x N x P]}, where \code{N} are the actors (senders in the array \code{rate} receivers in the array \code{choice})}

\item{actors}{vector of actors that may be observed interacting in the network. If \code{NULL}, actor names will be taken from the input edgelist.}

\item{directed}{logical value indicating whether dyadic events are directed (\code{TRUE}) or undirected (\code{FALSE}).}

\item{method}{optimization method to be used: \code{MLE}, \code{GD}, \code{GDADAM}, \code{BSIR}, \code{HMC}}

\item{model}{can be "tie" or "actor" oriented modeling. This argument plays a fundamental role when \code{omit_dyad} is supplied. Indeed, when actor-oriented modeling, the dynamic risk set will consist of two risk sets objects (senders' and dyads' risk sets). In the tie-oriented model the function will return a dynamic risk set referred at a dyad-level.}

\item{ncores}{number of threads for the parallelization (\code{default = 1}, no parallelization)}
}
\value{
list with fitted rems for each window
}
\description{
Function that fits a moving window relational event model (using remstimate).
}
