\name{REHdynamics-package}
\alias{REHdynamics-package}
\alias{REHdynamics}
\docType{package}
\title{
\packageTitle{REHdynamics}
}
\description{
\packageDescription{REHdynamics}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{REHdynamics}
\packageIndices{REHdynamics}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{REHdynamics}

Maintainer: \packageMaintainer{REHdynamics}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
