% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_constant.R
\docType{data}
\name{static_constant}
\alias{static_constant}
\title{static_constant}
\format{
A list of length 200, each element a remstimate object with the fit.
}
\usage{
static_constant
}
\description{
Results for the "static" REM for the 200 generated datasets in the constant 
effects scenario.
}
\keyword{datasets}
