% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_cyclic.R
\docType{data}
\name{static_cyclic}
\alias{static_cyclic}
\title{static_cyclic}
\format{
A list of length 200, each element a remstimate object with the fit.
}
\usage{
static_cyclic
}
\description{
Results for the "static" REM for the 200 generated datasets in the cyclic 
effects scenario.
}
\keyword{datasets}
