% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_gradual.R
\docType{data}
\name{static_gradual}
\alias{static_gradual}
\title{static_gradual}
\format{
A list of length 200, each element a remstimate object with the fit.
}
\usage{
static_gradual
}
\description{
Results for the "static" REM for the 200 generated datasets in the gradual 
effects scenario.
}
\keyword{datasets}
