% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_mixed.R
\docType{data}
\name{static_mixed}
\alias{static_mixed}
\title{static_mixed}
\format{
A list of length 200, each element a remstimate object with the fit.
}
\usage{
static_mixed
}
\description{
Results for the "static" REM for the 200 generated datasets in the mixed 
effects scenario.
}
\keyword{datasets}
